+++ src/Entities/Collection.php
@@ -19,6 +19,26 @@ class Collection extends NestedCollection
         return $this->injectValues();
     }
 
+    /**
+     * Load the fields of a belongs to or belongs to many entity relation
+     *
+     * @param  string $relation
+     * @param  mixed  $fields  Single field name string or array of field names
+     * @return $this
+     */
+    public function loadRelationFields($relation, $fields = ['*'])
+    {
+        if ( ! key_exists($relation, $this->first()->getAttributes())) {
+            $this->loadFields($relation);
+        }
+        $relations = $this->pluck($relation)->filter();
+        if ( ! $relations->first() instanceof Entity) {
+            $relations = $relations->flatten();
+        }
+        (new static($relations->unique()))->loadFields($fields);
+        return $this;
+    }
+
     public function injectValues()
     {
         foreach ($this->items as $item) {